requirements.txt
pystray
Pillow
watchdog
